Create ReadME
This example branch is for the use of pix2pix segmentation. The images need to have their full color background, and the masks need to be completely opaque in order to train. The other example branches i.e COCO does not do this. I am proposing altering the code to generate this type of data.